Move functions for checking if ime is enabled and selected into their own util class to reduce the size of the florisboard class
